notes_be/requirements.txt
Django
djangorestframework-simplejwt
djangorestframework
django-cors-headers